# openapi.yaml
openapi: 3.0.3
info:
  title: STD24064
  description: This is a specification of STD24064
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Serveur local


paths:
  /ping:
    get:
      summary: Vérifier l'état du serveur
      responses:
        '200':
          description: Serveur en ligne
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /cars:
    get:
      summary: Récupérer la liste des voitures
      responses:
        '200':
          description: Liste des voitures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Car'

    post:
      summary: Créer une nouvelle voiture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Car'
      responses:
        '201':
          description: Voiture créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'

  /cars/{id}:
    get:
      summary: Récupérer une voiture par ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Voiture trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '404':
          description: Voiture non trouvée

  /cars/{id}/characteristics:
    put:
      summary: Mettre à jour les caractéristiques d'une voiture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Characteristics'
      responses:
        '200':
          description: Voiture mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '404':
          description: Voiture non trouvée

components:
  schemas:
    Car:
      type: object
      properties:
        id:
          type: string
        brand:
          type: string
        model:
          type: string
        characteristics:
          $ref: '#/components/schemas/Characteristics'
    Characteristics:
      type: object
      properties:
        max_speed:
          type: number
          format: float
        max_fuel_capacity:
          type: number
          format: float